Update songShuffleV2_5
Added the remaining genres, will change comments to be more informative later, and finished the assignment. This code is easy and not optimized on purpose, because it was used to show my classmates a tutorial, and made to ensure that very new learners could ( somewhat ) understand it. This is probably not what the YouCubed curriculum intended(?)